refactor: collapse app.ts into main.ts and deepen runtime configuration
