+++ src/main.ts
@@ -1,20 +1,34 @@
 import { pathToFileURL } from "node:url"
 
 import { NodeRuntime } from "@effect/platform-node"
-import { Effect, Option } from "effect"
+import { Effect, Layer } from "effect"
 
-import { runApplication } from "./app.ts"
 import {
-  ActualConfigService,
-  Email2FAConfigService,
-  FintualConfigService,
   ScheduleConfigService,
   type RuntimeConfigError,
-  redactionSecrets,
   resolveRuntimeConfig,
+  runtimeLayer,
 } from "./env.ts"
 import { Job, type JobError } from "./job.ts"
 import { RedactingLogger, reportUnhandledFailure } from "./logging.ts"
+import { runScheduler } from "./scheduler.ts"
+
+export const runApplication = Effect.fn("Main.runApplication")(function* (): Effect.fn.Return<
+  void,
+  JobError,
+  ScheduleConfigService | Job
+> {
+  const schedule = yield* ScheduleConfigService
+  const job = yield* Job
+
+  if (schedule.mode === "schedule") {
+    return yield* runScheduler(job.synchronize(), schedule)
+  }
+
+  yield* Effect.logInfo("Running task once...")
+  yield* job.synchronize()
+  yield* Effect.logInfo("Task completed.")
+})
 
 const main = Effect.fn("Main.main")(function* (): Effect.fn.Return<
   void,
@@ -28,15 +42,7 @@ const main = Effect.fn("Main.main")(function* (): Effect.fn.Return<
 
   yield* runApplication().pipe(
     Effect.tapCause(reportUnhandledFailure),
-    Effect.provide(Job.live),
-    Effect.provideService(ActualConfigService, runtimeConfig.actual),
-    Effect.provideService(FintualConfigService, runtimeConfig.fintual),
-    Effect.provideService(
-      Email2FAConfigService,
-      runtimeConfig.email2FA ? Option.some(runtimeConfig.email2FA) : Option.none(),
-    ),
-    Effect.provideService(ScheduleConfigService, runtimeConfig.schedule),
-    Effect.provide(RedactingLogger.layer(redactionSecrets(runtimeConfig))),
+    Effect.provide(Job.live.pipe(Layer.provideMerge(runtimeLayer(runtimeConfig)))),
   )
 })
